+++ lib/Service/CospendService.php
@@ -948,7 +948,7 @@ class CospendService {
 	 * @param string $userId
 	 * @param array $projectInfo
 	 * @param array $bills
-	 * @param string|null $name
+	 * @param string|null $outputPath
 	 * @return array
 	 * @throws InvalidPathException
 	 * @throws LockedException
@@ -956,39 +956,55 @@ class CospendService {
 	 * @throws NotFoundException
 	 * @throws NotPermittedException
 	 */
-	public function exportCsvProject(string $projectId, string $userId, array $projectInfo, array $bills, ?string $name = null): array {
-		// create export directory if needed
-		$outPath = $this->userConfig->getValueString($userId, Application::APP_ID, 'outputDirectory', '/Cospend', lazy: true);
-		$userFolder = $this->root->getUserFolder($userId);
-		$msg = $this->createAndCheckExportDirectory($userFolder, $outPath);
-		if ($msg !== '') {
-			return ['message' => $msg];
-		}
-		$folder = $userFolder->get($outPath);
-		if (!$folder instanceof Folder) {
-			return ['message' => $outPath . ' is not a directory'];
-		}
-
+	public function exportCsvProject(
+		string $projectId, string $userId, array $projectInfo, array $bills,
+		?string $outputPath = null, bool $fileSystem = false,
+	): array {
 		// create file
-		$filename = $projectId . '.csv';
-		if ($name !== null) {
-			$filename = $name;
-			if (!str_ends_with($filename, '.csv')) {
-				$filename .= '.csv';
+		if ($outputPath === null) {
+			$outputPath = $projectId . '.csv';
+		} else {
+			if (str_contains($outputPath, '..')) {
+				throw new \Exception('Invalid file name or path');
+			}
+			if (!str_ends_with($outputPath, '.csv')) {
+				$outputPath .= '.csv';
 			}
 		}
-		if ($folder->nodeExists($filename)) {
-			$folder->get($filename)->delete();
+		if ($fileSystem) {
+			$handler = fopen($outputPath, 'w');
+			if ($handler === false) {
+				return ['message' => 'Failed to open file "' . $outputPath . '" for writing'];
+			}
+		} else {
+			// create export directory if needed
+			$userFolder = $this->root->getUserFolder($userId);
+			$outputDirPath = $this->userConfig->getValueString($userId, Application::APP_ID, 'outputDirectory', '/Cospend', lazy: true);
+			$msg = $this->createAndCheckExportDirectory($userFolder, $outputDirPath);
+			if ($msg !== '') {
+				return ['message' => $msg];
+			}
+			$folder = $userFolder->get($outputDirPath);
+			if (!$folder instanceof Folder) {
+				return ['message' => $outputDirPath . ' is not a directory'];
+			}
+			if ($folder->nodeExists($outputPath)) {
+				$folder->get($outputPath)->delete();
+			}
+			$storageFile = $folder->newFile($outputPath);
+			$handler = $storageFile->fopen('w');
 		}
-		$file = $folder->newFile($filename);
-		$handler = $file->fopen('w');
 		foreach ($this->getJsonProject($projectInfo, $bills) as $chunk) {
 			fwrite($handler, $chunk);
 		}
 
 		fclose($handler);
-		$file->touch();
-		return ['path' => $outPath . '/' . $filename];
+		if ($fileSystem) {
+			return ['path' => $outputPath];
+		} else {
+			$storageFile->touch();
+			return ['path' => $outputDirPath . '/' . $outputPath];
+		}
 	}
 
 	/**